refactor: Make ItemType an enum.

+++ VisualPinball.Engine/VPT/HitTarget/HitTargetData.cs
@@ -119,7 +119,7 @@ namespace VisualPinball.Engine.VPT.HitTarget
 
 		public override void Write(BinaryWriter writer, HashWriter hashWriter)
 		{
-			writer.Write(ItemType.HitTarget);
+			writer.Write((int)ItemType.HitTarget);
 			Write(writer, Attributes, hashWriter);
 			WriteEnd(writer, hashWriter);
 		}
