+++ VisualPinball.Engine/VPT/Kicker/KickerData.cs
@@ -79,7 +79,7 @@ namespace VisualPinball.Engine.VPT.Kicker
 
 		public override void Write(BinaryWriter writer, HashWriter hashWriter)
 		{
-			writer.Write(ItemType.Kicker);
+			writer.Write((int)ItemType.Kicker);
 			Write(writer, Attributes, hashWriter);
 			WriteEnd(writer, hashWriter);
 		}
